Align import statement
When doing `Organize imports` (if configured with `haskutil.alignImports`)

+++ src/features/sortImportProvider.ts
@@ -30,28 +30,69 @@ export default class SortImportProvider implements CodeActionProvider
     this.diagnosticCollection.dispose();
     this.command.dispose();
   }
+
+  public static ensureAligned(newImport: ImportDeclaration, existingImports: ImportDeclaration[]): ImportDeclaration
+  {
+    if (SortImportProvider.shouldAlignImports)
+    {
+      let allImports = existingImports.map(imp => imp);
+      allImports.unshift(newImport);
+      SortImportProvider.alignImports(allImports);
+    }
+    return newImport;
+  }
+
+  private static get shouldAlignImports(): boolean
+  {
+    var extSettings = vscode.workspace.getConfiguration("haskutil");
+    return extSettings.get("alignImports");
+  }
   
   private checkImports(document: TextDocument)
   {
     const imports = ImportDeclaration.getImports(document.getText());
+    let messages = [];
+    
+    const aligned =
+      imports.length === 0 || (
+        imports.some(imp => imp.qualified.trim() === "qualified") &&
+        imports.every(imp => imp.qualified.length === " qualified ".length)
+      ) || (
+        imports.every(imp => imp.qualified.trim() === "") &&
+        imports.every(imp => imp.qualified.length === " ".length)
+      );
+    if (!aligned && SortImportProvider.shouldAlignImports)
+    {
+      messages = ["not aligned"];
+    }
+
     let pred = "";
     for (const imp of imports)
     {
-      if (imp.module < pred)
+      const curr = imp.module + (imp.importList || "");
+      if (curr < pred)
       {
-        const lastImport = imports[imports.length - 1];
-        const range = new Range(
-          document.positionAt(imports[0].offset),
-          document.positionAt(lastImport.offset + lastImport.length));
-        const message = "Unsorted imports";
-        const diagnostic = new Diagnostic(range, message, DiagnosticSeverity.Hint);
-        diagnostic.code = SortImportProvider.diagnosticCode;
-        this.diagnosticCollection.set(document.uri, [diagnostic]);
-        return;
+        messages.unshift("unsorted");
+        break;
       }
-      pred = imp.module;
+      pred = curr;
     }
-    this.diagnosticCollection.delete(document.uri);
+
+    if (messages.length > 0)
+    {
+      const lastImport = imports[imports.length - 1];
+      const range = new Range(
+        document.positionAt(imports[0].offset),
+        document.positionAt(lastImport.offset + lastImport.length));
+      const message = `Imports are ${messages.join(" and ")}`;
+      const diagnostic = new Diagnostic(range, message, DiagnosticSeverity.Hint);
+      diagnostic.code = SortImportProvider.diagnosticCode;
+      this.diagnosticCollection.set(document.uri, [diagnostic]);
+    }
+    else
+    {
+      this.diagnosticCollection.delete(document.uri);
+    }
   } 
 
 	public async provideCodeActions(document: TextDocument, range: Range, context: CodeActionContext, token: CancellationToken): Promise<CodeAction[]>
@@ -59,7 +100,7 @@ export default class SortImportProvider implements CodeActionProvider
 		let codeActions = [];
     for (let diagnostic of context.diagnostics.filter(d => d.code === SortImportProvider.diagnosticCode))
     {
-      let title = "Sort imports";
+      let title = "Organize imports";
       let codeAction = new CodeAction(title, CodeActionKind.QuickFix);
       codeAction.command = {
         title: title,
@@ -77,7 +118,14 @@ export default class SortImportProvider implements CodeActionProvider
 	private runCodeAction(document: TextDocument): Thenable<boolean>
   {
     const oldImports = ImportDeclaration.getImports(document.getText());
-    const newImports = oldImports.map(i => i).sort((l, r) => l.module.localeCompare(r.module));
+    let newImports = oldImports
+      .map(i => i)
+      .sort((l, r) => (l.module + (l.importList || "")).localeCompare(r.module + (r.importList || "")));
+    if (SortImportProvider.shouldAlignImports)
+    {
+      newImports = SortImportProvider.alignImports(newImports);
+    }
+
     var edit = new WorkspaceEdit();
     for (let i = oldImports.length - 1; i >= 0; i--)
     {
@@ -85,5 +133,29 @@ export default class SortImportProvider implements CodeActionProvider
       edit.insert(document.uri, oldImports[i].getRange(document).start, newImports[i].text);
     }
     return vscode.workspace.applyEdit(edit);
-	}
+  }
+  
+  private static alignImports(imports: ImportDeclaration[]): ImportDeclaration[]
+  {
+    const isQualified = imp => imp.qualified.trim() === "qualified";
+
+    return imports.some(isQualified) ?
+      imports.map(imp =>
+      {
+        if (isQualified(imp))
+        {
+          imp.qualified = " qualified ";
+        }
+        else
+        {
+          imp.qualified = "           ";
+        }
+        return imp;
+      }) :
+      imports.map(imp =>
+      {
+        imp.qualified = " ";
+        return imp;
+      });
+  }
 }
